app.js updated
